fix(menu, filters): 'show only favorites' indication in the filters menu badge #678
